Удаляет ненужные функции академии, исправляет модуляцию
